gene prioritisation component tabs
