+++ apps/genetics/src/sections/studyLocus/GenePrioritisationTabs/index.js
@@ -1,0 +1,1 @@
+export { default } from './GenePrioritisationTabs';
